+++ client/package.json
@@ -10,6 +10,7 @@
     "papaparse": "^5.3.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "^3.0.1",
     "tabletop": "^1.6.3",
     "web-vitals": "^2.1.3",
