--- client/src/components/features/FormSearch.js
@@ -1,19 +1,0 @@
-import React from 'react';
-
-const FormSearch = (props) => {
-    return ( 
-    <form action="">
-        <label htmlFor="serch">Szukaj : 
-            <input
-                name="search"
-                type="text" 
-                id="search"
-                onChange={props.change}
-                value={props.serchText}
-            />
-        </label>
-    </form>  
-    );
-}
- 
-export default FormSearch;
