+++ client/src/components/layout/MainLayot.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const MainLayout = ({ children }) => {
+  return <div>{children}</div>;
+};
+
+export default MainLayout;
