+++ client/src/components/pages/HomePage.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+const HomePage = () => {
+  return (
+    <div>
+      <h1>Witam !!!</h1>
+    </div>
+  );
+};
+export default HomePage;
