+++ client/src/components/pages/NotFoundPage.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+const NotFoundPage = () => {
+  return (
+    <div>
+      <h1>404 Not Found</h1>
+    </div>
+  );
+};
+export default NotFoundPage;
